+++ src/main.cpp
@@ -22,7 +22,7 @@
 #include "RotarySwitch.h"
 
 #include "projectConfig.h"
-
+#include "effectsConfig.h"
 
 
 esp_event_loop_handle_t loop_handle;
@@ -207,34 +207,6 @@ void modeChangeToLocalCallback(void* hander_arg, esp_event_base_t event_base, in
 }
 
 
-
-/**
- * @brief read values from gyro sensor
- * 
- * @param[in] rxvalues create int16_t rxvalues[7] before and pass pointer to it. will be filled with values for accelx, accely accelz, temp, gyrox, gyroy, gyroz
- */
-void readGyro(int16_t *rxvalues) {
-    static const uint8_t startregister = 0x3B;
-    uint8_t rx_data[14];
-    ESP_ERROR_CHECK_WITHOUT_ABORT(i2c_master_write_read_device(I2C_NUM_0, conf::I2C_GYRO_ADDR, &startregister, sizeof(startregister), rx_data, sizeof(rx_data), conf::I2C_TIMEOUT_MS/portTICK_RATE_MS));
-    //ESP_LOG_BUFFER_HEX_LEVEL("i2c_gyro_rx", rx_data, sizeof(rx_data), ESP_LOG_VERBOSE);
-    for (int i = 0; i < 7; i++,rxvalues++) {
-        *rxvalues = htons(*((int16_t*) rx_data+i));
-    }
-    ESP_LOG_LEVEL(ESP_LOG_VERBOSE, "readGyro", "accelx: %7d, accely: %7d, accelz: %7d, temp: %7d, gyrox: %7d, gyroy: %7d, gyroz: %7d\r", (rxvalues[0]), (rxvalues[1]), (rxvalues[2]), (rxvalues[3]), (rxvalues[4]), (rxvalues[5]), (rxvalues[6]));
-    //rxvalues = reinterpret_cast<int16_t*>(rx_data);
-}
-
-/**
- * @return Norm of the acceleration vector if acceleration in x direction is positive, negative norm oterwise.
- */
-float directedAccelNorm() {
-    int16_t rxvalues[7];
-    readGyro(rxvalues);
-    const int sign = rxvalues[0] < 0 ? (-1) : 1;
-    return sign * sqrtf(pow(rxvalues[0], 2) + pow(rxvalues[1], 2) + pow(rxvalues[2], 2));
-}
-
 /**
  * @brief let the onboard led sine-fade
  */
@@ -265,178 +237,6 @@ void ledFun(void*) {
         ESP_ERROR_CHECK(ledc_update_duty(LEDC_HIGH_SPEED_MODE, LEDC_CHANNEL_0));
         val+=100;
         vTaskDelay(1);
-    }
-}
-
-DEFINE_GRADIENT_PALETTE(heatmap_fire_) {
-           0,   0,   0,   0, // black
-         128, 255,   0,   0,   //red
-         224, 255, 255,   0,   //bright yellow
-         255, 255, 255, 255, // white
-    };
-
-DEFINE_GRADIENT_PALETTE(heatmap_test_) {
-           0,   0, 255,   0, // blue
-         255, 255,   0,   0, // yellow
-};
-
-DEFINE_GRADIENT_PALETTE(heatmap_tooff_) {
-           0,   0, 255,   0, // blue
-         255,   0,   0,   0, // yellow
-};
-
-
-/**
- * @brief handle local tasks
- * 
- */
-void localtask(void*) {
-    ESP_LOG_LEVEL(ESP_LOG_DEBUG, __FUNCTION__, "%s started", __FUNCTION__);
-    // case 0
-    int16_t rxvalues[7];
-    static uint8_t r, g, b;
-    const float alpha = 1.0/16.0;
-
-    // case 6
-    unsigned long time_now, time_prev;
-    time_now = time_prev = millis();
-    float norm_accel_max = 0, norm_reading = 0, norm_prev = 0;
-    static float norm_accel = 0;
-    CRGBPalette16 fire = heatmap_fire_;
-    CRGBPalette16 test = heatmap_test_;
-    CRGBPalette16 tooff = heatmap_tooff_;
-    float jerk = 0;
-    static float jerk_prev = 0, jerk_max = 0, jerk_min = 0;
-    float directednormaccel = 0;
-    uint8_t brightness = 0;
-
-    
-
-
-    while (1) {
-        ESP_LOG_LEVEL(ESP_LOG_VERBOSE, __FUNCTION__, "high watermark: %d, rotaryswitch: %d", uxTaskGetStackHighWaterMark(NULL), rotaryswitch);
-        FastLED.show(conf::MAX_BRIGHTNESS);
-        /**
-         * @todo 
-         * line @lineinfo: replace  switch with array of function pointers and
-         * put array in projectConfig.h
-         */
-        switch (rotaryswitch) {
-        case 0:
-            readGyro(rxvalues);
-            r = ((1-alpha) * r ) + (alpha * (abs(rxvalues[0]) >> 7));
-            g = ((1-alpha) * g ) + (alpha * (abs(rxvalues[1]) >> 7));
-            b = ((1-alpha) * b ) + (alpha * (abs(rxvalues[2]) >> 7));
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB(r, g, b));
-            //FastLED.show();
-            break;
-        case 1:
-            /* ---------------------------------- jerk ---------------------------------- */
-            time_now = millis();
-            norm_reading = directedAccelNorm();
-            jerk = (norm_reading - norm_prev)/(time_now-time_prev);
-            time_prev = time_now;
-            norm_prev = norm_reading;
-            // bounds
-            jerk_max = jerk_max * 0.99;
-            jerk_min = jerk_min * 0.99;
-            if (jerk > jerk_max) jerk_max = jerk;
-            else if (jerk < jerk_min) jerk_min = jerk;
-            ESP_LOG_LEVEL(ESP_LOG_DEBUG, __FUNCTION__, "jerk: %f", jerk);
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::White);
-            // FastLED.show(mapfb(jerk, jerk_min, jerk_max, 0, 255));
-            //FastLED.show(255 * (jerk > 100.0));
-            break;
-        case 2:
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Green);
-            readGyro(rxvalues);
-            ESP_LOG_LEVEL(ESP_LOG_INFO, "readGyro", "accelx: %7d, accely: %7d, accelz: %7d, temp: %7d, gyrox: %7d, gyroy: %7d, gyroz: %7d\r", (rxvalues[0]), (rxvalues[1]), (rxvalues[2]), (rxvalues[3]), (rxvalues[4]), (rxvalues[5]), (rxvalues[6]));
-
-            //FastLED.show(conf::MAX_BRIGHTNESS);
-            //vTaskSuspend(NULL);
-            break;
-        case 3:
-            directednormaccel = directedAccelNorm();
-            if (directednormaccel < -10000) {
-                fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Green);
-            } else {
-                fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Black);
-            }
-            ESP_LOG_LEVEL(ESP_LOG_INFO, "accel", "acceleration norm: %f\r", directednormaccel);
-            //FastLED.show(conf::MAX_BRIGHTNESS);
-            //vTaskSuspend(NULL);
-            break;
-        case 4:
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Lime);
-            //FastLED.show(conf::MAX_BRIGHTNESS);
-            //vTaskSuspend(NULL);
-            break;
-        case 5:
-            readGyro(rxvalues);
-            
-            norm_reading = sqrtf(pow(rxvalues[0], 2) + pow(rxvalues[1], 2) + pow(rxvalues[2], 2)) - conf::G_VAL;
-            if (norm_accel_max < norm_accel) 
-                norm_accel_max = norm_accel;
-            if (norm_reading < norm_accel + 200) {
-                norm_accel = (norm_reading * alpha) +  (norm_accel * (1.0 - alpha));
-            } else {
-                norm_accel = norm_reading;
-            }
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, ColorFromPalette(tooff, 0xff * (norm_accel/norm_accel_max)));
-            //FastLED.show();
-            break;
-        case 6:
-            readGyro(rxvalues);
-            
-            norm_reading = sqrtf(pow(rxvalues[0], 2) + pow(rxvalues[1], 2) + pow(rxvalues[2], 2)) - conf::G_VAL;
-            if (norm_accel_max < norm_accel) 
-                norm_accel_max = norm_accel;
-            if (norm_reading < norm_accel + 200) {
-                norm_accel = (norm_reading * alpha) +  (norm_accel * (1.0 - alpha));
-            } else {
-                norm_accel = norm_reading;
-            }
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, ColorFromPalette(fire, 0xff * (norm_accel/norm_accel_max)));
-            //FastLED.show();
-            break;
-        case 7:
-            readGyro(rxvalues);
-            ESP_LOG_LEVEL(ESP_LOG_INFO, "readGyro", "accelx: %7d, accely: %7d, accelz: %7d, temp: %7d, gyrox: %7d, gyroy: %7d, gyroz: %7d\r", (rxvalues[0]), (rxvalues[1]), (rxvalues[2]), (rxvalues[3]), (rxvalues[4]), (rxvalues[5]), (rxvalues[6]));
-            directednormaccel = directedAccelNorm();
-            if (directednormaccel < -10000) {
-                brightness = conf::MAX_BRIGHTNESS;
-                fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Cyan);
-            } else {
-                brightness = brightness * 0.95;
-            }
-            ESP_LOG_LEVEL(ESP_LOG_INFO, "accel", "acceleration norm: %f\r", directednormaccel);
-            //FastLED.show();
-            //vTaskSuspend(NULL);
-            break;
-        
-        case 8:
-            readGyro(rxvalues);
-            
-            norm_reading = sqrtf(pow(rxvalues[0], 2) + pow(rxvalues[1], 2) + pow(rxvalues[2], 2)) - conf::G_VAL;
-            if (norm_accel_max < norm_accel) 
-                norm_accel_max = norm_accel;
-            if (norm_reading < norm_accel + 200) {
-                norm_accel = (norm_reading * alpha) +  (norm_accel * (1.0 - alpha));
-            } else {
-                norm_accel = norm_reading;
-            }
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, ColorFromPalette(test, 0xff * (norm_accel/norm_accel_max)));
-            
-            break;
-
-        
-        default:
-            fill_solid(*leds, conf::NUM_LEDS_TOTAL, CRGB::Black);
-            //FastLED.show();
-            //vTaskSuspend(NULL);
-            break;
-        }
-        vTaskDelay(5);
     }
 }
 
@@ -746,22 +546,3 @@ void otaFun(void*) {
 }
 
 
-/**
- * @brief scales a float variable in [in_min, in_max] linearly to [out_min, out_max]
- * 
- * @param in input value
- * @param in_min input lower bound
- * @param in_max input upper bound
- * @param out_min output lower bound
- * @param out_max output upper bound
- * @retval out_min if `in` <= `in_min`
- * @retval out_max if `in` >= `in_max`
- * @return mapped value
- */
-float mapfb(float in, float in_min, float in_max, float out_min, float out_max) {
-    if (in <= in_min)
-        return out_min;
-    if (in_min >= in_max)
-        return out_max;
-    return (in - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
-}
